backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"

dependencies = [
    "fastapi[standard]>=0.115.6",
    "mongoengine>=0.29.1",
    "motor>=3.6.0",
    "passlib>=1.7.4",
    "pydantic>=2.10.5",
    "pyjwt>=2.10.1",
    "pyotp>=2.9.0",
    "python-dotenv>=1.0.1",
    "starlette>=0.41.3",
]

